+++ Assets/Scripts/Interfaces/IDamagable.cs
@@ -10,5 +10,5 @@ public interface IDamagable
     void ModifyHealth(int amount);
 
     int knockbackGiven { get; set; }
-    void KnockbackOnDamage(int amount, int dirX, int dirY);
+    void KnockbackOnDamage(int amount, float dirX, float dirY);
 }
